+++ sconnConnector/POCO/Config/sconn/IO/Output/sconnOutput.cs
@@ -25,12 +25,12 @@ namespace sconnConnector.POCO.Config
     public class sconnOutput : IAlarmSystemConfigurationEntity, ISerializableConfiguration, IFakeAbleConfiguration
     {
 
-        public int Id { get; set; }
+        public byte Id { get; set; }
         public sconnOutputType Type { get; set; }
         public bool Value { get; set; }
         public bool Enabled { get; set; }
         public string Name { get; set; }
-        public int NameId { get; set; }
+        public byte NameId { get; set; }
         private static Logger _logger = LogManager.GetCurrentClassLogger();
 
         public sconnOutput()
